Split main module in to smaller files
