+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "graphql2rest",
-	"version": "0.6.3",
+	"version": "0.6.4",
 	"author": "Roy Mor <roy.mor@sisense.com>",
 	"contributors": [
 		"Ray Luxembourg",
